review.tsv updates - Refactor: Moved review.tsv code into a separate function in order to make the disease2gene section more readable.
